requirements.txt
PyYAML==5.4.1
firebase-admin==4.5.3
requests==2.25.1
beautifulsoup4==4.12.3
openai==1.13.3
pytz==2024.1
firebase_admin==4.5.3
